Update mod.rs
I've added some tests for this file. Checks to make sure you can add a player, add a deck, add more than one deck, remove a deck, and checks to see if the player is registered or not. Hopefully I covered everything. Because my IDE is broken I haven't checked to see which of these tests work or not.