Update tournament.rs
I've removed the need for TOs to prune decks. Instead, if a player tries to add a deck after the max number is exceeded (default now set to 1) then they get an error. I'm not sure if InvalidDeckCount is the right error message to use though, as that's the same message a TO would get if they try to set the min deck to greater than max or max less than min. 

I'm pretty sure this also means that the PruneDecks stuff from admin_ops.rs can be removed as well. But because I suck at using GitHub I couldn't figure out how to also attach that change to this branch.